Pass knob settings down via a PageContext
Currently that's isInstructorView and maxAttempts. These likely *will* have to
be props when we actually implement this in Core (since we don't have page-wide
contexts available there due to how react-rails works), but we'll figure that
out later. (Not sure if we'd pass just maxAttempts, or if we need to pass the
entire PageRespondablePolicy to each QP component.)